Refactor PAF output code
