+++ bin/.pylintrc
@@ -1,5 +1,5 @@
 [basic]
-good-names = i, mx, j, a, b, k, n, d, s, t, e, m, cr
+good-names = i, mx, j, a, b, k, n, d, s, t, e, m, cr, i1, i2
 max-line-length = 120
 max-args = 10
 max-locals = 50
